Added tasks header file, changed order within headers
