+++ src/dd_tasks.h
@@ -1,0 +1,24 @@
+#ifndef DD_TASKS_H
+#define DD_TASKS_H
+
+/* Standard includes. */
+#include <stdint.h>
+#include <stdio.h>
+#include "../FreeRTOS_Source/include/FreeRTOS.h"
+#include "../FreeRTOS_Source/include/queue.h"
+#include "../FreeRTOS_Source/include/semphr.h"
+#include "../FreeRTOS_Source/include/task.h"
+#include "../FreeRTOS_Source/include/timers.h"
+
+// enum for state
+enum task_type {PERIODIC,APERIODIC};
+
+// deadline-driven schedule handler struct
+struct dd_task {
+    TaskHandle_t t_handle;
+    task_type type;
+    uint32_t task_id;
+    uint32_t release_time;
+    uint32_t absolute_deadline;
+    uint32_t completion_time;
+};
